+++ packages/terra-time-input/CHANGELOG.md
@@ -2,10 +2,13 @@
 
 ## Unreleased
 
+* Fixed
+  * Accessibility improvements.
+
 ## 4.44.1 - (February 9, 2022)
 
 * Added
-  * Added a test for time-input padding. 
+  * Added a test for time-input padding.
 * Fixed
   * Reduce random test failures.